Restructured databases folder
