+++ cogs/archiveuse/profile_manager.py
@@ -1,12 +1,12 @@
 import traceback
 import discord
 from libs.config import Config
-from libs.databases.profile_layout.profile_layout import ProfileLayout
-from libs.databases.profile_layout.profile_layouts import ProfileLayouts
+from libs.databases.model.profile_layout.profile_layout import ProfileLayout
+from libs.databases.model.profile_layout.profile_layouts import ProfileLayouts
 
-from libs.databases.user.user import User
-from libs.databases.wallpaper.wallpaper import Wallpaper
-from libs.databases.wallpaper.wallpapers import Wallpapers
+from libs.databases.model.user.user import User
+from libs.databases.model.wallpaper.wallpaper import Wallpaper
+from libs.databases.model.wallpaper.wallpapers import Wallpapers
 from libs.exception.color.color_not_correct_exception import ColorNotCorrectException
 from libs.exception.handler import Handler
 from libs.exception.smartpoint.not_enougt_smartpoint_exception import NotEnougtSmartpointException
